remove memory backend
