--- internal/testutil/prefix.go
@@ -1,127 +1,0 @@
-package testutil
-
-import (
-	"strings"
-
-	"github.com/SiaFoundation/s3d/s3"
-)
-
-// newPrefix creates a new Prefix from the given optional prefix and delimiter
-// strings.
-func newPrefix(prefix, delim *string) (p s3.Prefix) {
-	if prefix != nil {
-		p.HasPrefix, p.Prefix = true, *prefix
-	}
-	if delim != nil {
-		p.HasDelimiter, p.Delimiter = true, *delim
-	}
-	return p
-}
-
-// filePrefix returns the path portion, then the remaining portion of the
-// Prefix if the Delimiter is "/". If the Delimiter is not set, or not "/",
-// ok will be false.
-//
-// For example:
-//
-//	/foo/bar/  : path: /foo/bar  remaining: ""
-//	/foo/bar/b : path: /foo/bar  remaining: "b"
-//	/foo/bar   : path: /foo      remaining: "bar"
-func filePrefix(p s3.Prefix) (path, remaining string, ok bool) {
-	if !p.HasPrefix || !p.HasDelimiter || p.Delimiter != "/" {
-		return "", "", p.Delimiter == "/"
-	}
-
-	idx := strings.LastIndexByte(p.Prefix, '/')
-	if idx < 0 {
-		return "", p.Prefix, true
-	} else {
-		return p.Prefix[:idx], p.Prefix[idx+1:], true
-	}
-}
-
-// match checks whether key starts with a prefix. If the prefix does not match,
-// nil is returned.
-//
-// It is a best-effort attempt to implement the prefix/delimiter matching found
-// in S3.
-//
-// To check whether the key belongs in Contents or CommonPrefixes, compare the
-// result to key.
-func match(p s3.Prefix, key string) *prefixMatch {
-	if !p.HasPrefix && !p.HasDelimiter {
-		// If there is no prefix in the search, the match is the prefix:
-		return &prefixMatch{Key: key, MatchedPart: key}
-	}
-
-	if !p.HasDelimiter {
-		// If the request does not contain a delimiter, prefix matching is a
-		// simple string prefix:
-		if strings.HasPrefix(key, p.Prefix) {
-			return &prefixMatch{Key: key, MatchedPart: p.Prefix}
-		}
-		return nil
-	}
-
-	// Delimited + Prefix matches, for example:
-	//	 $ aws s3 ls s3://my-bucket/
-	//	                            PRE AWSLogs/
-	//	 $ aws s3 ls s3://my-bucket/AWSLogs
-	//	                            PRE AWSLogs/
-	//	 $ aws s3 ls s3://my-bucket/AWSLogs/
-	//	                            PRE 260839334643/
-	//	 $ aws s3 ls s3://my-bucket/AWSLogs/2608
-	//	                            PRE 260839334643/
-
-	keyParts := strings.Split(strings.TrimLeft(key, p.Delimiter), p.Delimiter)
-	preParts := strings.Split(strings.TrimLeft(p.Prefix, p.Delimiter), p.Delimiter)
-
-	if len(keyParts) < len(preParts) {
-		return nil
-	}
-
-	// If the key exactly matches the prefix, but only up to a delimiter,
-	// AWS appends the delimiter to the result:
-	//	 $ aws s3 ls s3://my-bucket/AWSLogs
-	//	                            PRE AWSLogs/
-	appendDelim := len(keyParts) != len(preParts)
-	matched := 0
-
-	last := len(preParts) - 1
-	for i := range preParts {
-		if i == last {
-			if !strings.HasPrefix(keyParts[i], preParts[i]) {
-				return nil
-			}
-		} else {
-			if keyParts[i] != preParts[i] {
-				return nil
-			}
-		}
-		matched++
-	}
-
-	if matched == 0 {
-		return nil
-	}
-
-	out := strings.Join(keyParts[:matched], p.Delimiter)
-	if appendDelim {
-		out += p.Delimiter
-	}
-	isCommonPrefix := out != key || (appendDelim && strings.HasSuffix(key, p.Delimiter))
-	return &prefixMatch{Key: key, CommonPrefix: isCommonPrefix, MatchedPart: out}
-}
-
-// prefixMatch describes a successful match of a key against a Prefix.
-type prefixMatch struct {
-	// Input key passed to PrefixMatch.
-	Key string
-
-	// CommonPrefix indicates whether this key should be returned in the bucket
-	// contents or the common prefixes part of the "list bucket" response.
-	CommonPrefix bool
-
-	// The longest matched part of the key.
-	MatchedPart string
-}
